Move ParseSortFromTerm to SearchHelper
