+++ GitHubExtension/Helpers/SearchHelper.cs
@@ -3,7 +3,9 @@
 // See the LICENSE file in the project root for more information.
 
 using System.Globalization;
+using System.Text.RegularExpressions;
 using GitHubExtension.DataModel.Enums;
+using Octokit;
 
 namespace GitHubExtension.Helpers;
 
@@ -168,6 +170,53 @@ public static class SearchHelper
         {
             return null;
         }
+    }
+
+    public static (IssueSearchSort SortField, SortDirection Direction, string UpdatedTerm)? ParseSortFromTerm(string term)
+    {
+        if (string.IsNullOrWhiteSpace(term))
+        {
+            return null;
+        }
+
+        // Regex to match sort:field-direction (e.g., sort:created-asc)
+        var match = Regex.Match(term, @"sort:(\w+)-(asc|desc)", RegexOptions.IgnoreCase);
+        if (!match.Success)
+        {
+            return null;
+        }
+
+        var field = match.Groups[1].Value.ToLowerInvariant();
+        var direction = match.Groups[2].Value.ToLowerInvariant();
+
+        IssueSearchSort sortField;
+
+        // reactions and interactions are not supported in the Octokit API
+        switch (field)
+        {
+            case "created":
+                sortField = IssueSearchSort.Created;
+                break;
+            case "updated":
+                sortField = IssueSearchSort.Updated;
+                break;
+            case "comments":
+                sortField = IssueSearchSort.Comments;
+                break;
+            default:
+                return null;
+        }
+
+        var order = direction == "asc" ? SortDirection.Ascending : SortDirection.Descending;
+
+        // Remove the sort:field-direction part from the term, preserving single spaces between terms
+        var updatedTerm = Regex.Replace(term, @"\s*sort:\w+-\w+\s*", " ", RegexOptions.IgnoreCase)
+            .Trim();
+
+        // Replace multiple spaces with a single space
+        updatedTerm = Regex.Replace(updatedTerm, @"\s{2,}", " ");
+
+        return (sortField, order, updatedTerm);
     }
 
     private static readonly Dictionary<string, SearchType> SearchTypeMappings = new()
